feat: implementing developers routes and their validations

+++ backend/app/Services/levelService.php
@@ -36,6 +36,10 @@ class LevelService
 
     public function deleteLevel(Level $level): bool
     {
+        if ($level->hasDeveloper()) {
+            return false;
+        }
+
         $level->delete();
         return true;
     }
